Improve rack middlewares
* Move from prometheus/client/rack/* to rack/middleware/*
* Use histogram instead of summary in collector middleware
* Spepicfy collector metrics by changing the prefix to http_server_*
* Remove HTTP Host header from default label builder
