+++ examples/rack/config.ru
@@ -1,9 +1,22 @@
 require 'rack'
-require 'prometheus/client/rack/collector'
-require 'prometheus/client/rack/exporter'
+require 'prometheus/middleware/collector'
+require 'prometheus/middleware/exporter'
 
 use Rack::Deflater, if: ->(_, _, _, body) { body.any? && body[0].length > 512 }
-use Prometheus::Client::Rack::Collector
-use Prometheus::Client::Rack::Exporter
+use Prometheus::Middleware::Collector
+use Prometheus::Middleware::Exporter
 
-run ->(_) { [200, { 'Content-Type' => 'text/html' }, ['OK']] }
+srand
+
+app = lambda do |_|
+  case rand
+  when 0..0.8
+    [200, { 'Content-Type' => 'text/html' }, ['OK']]
+  when 0.8..0.95
+    [404, { 'Content-Type' => 'text/html' }, ['Not Found']]
+  else
+    raise NoMethodError, 'It is a bug!'
+  end
+end
+
+run app
