+++ spec/prometheus/middleware/collector_spec.rb
@@ -1,9 +1,9 @@
 # encoding: UTF-8
 
 require 'rack/test'
-require 'prometheus/client/rack/collector'
+require 'prometheus/middleware/collector'
 
-describe Prometheus::Client::Rack::Collector do
+describe Prometheus::Middleware::Collector do
   include Rack::Test::Methods
 
   let(:registry) do
@@ -26,7 +26,7 @@ describe Prometheus::Client::Rack::Collector do
   end
 
   it 'handles errors in the registry gracefully' do
-    counter = registry.get(:http_requests_total)
+    counter = registry.get(:http_server_requests_total)
     expect(counter).to receive(:increment).and_raise(NoMethodError)
 
     get '/foo'
@@ -36,16 +36,16 @@ describe Prometheus::Client::Rack::Collector do
 
   it 'traces request information' do
     expect(Time).to receive(:now).twice.and_return(0.0, 0.2)
-    labels = { method: 'get', host: 'example.org', path: '/foo', code: '200' }
 
     get '/foo'
 
-    {
-      http_requests_total: 1.0,
-      http_request_duration_seconds: { 0.5 => 0.2, 0.9 => 0.2, 0.99 => 0.2 },
-    }.each do |metric, result|
-      expect(registry.get(metric).get(labels)).to eql(result)
-    end
+    metric = :http_server_requests_total
+    labels = { method: 'get', path: '/foo', code: '200' }
+    expect(registry.get(metric).get(labels)).to eql(1.0)
+
+    metric = :http_server_request_duration_seconds
+    labels = { method: 'get', path: '/foo' }
+    expect(registry.get(metric).get(labels)).to include(0.1 => 0, 0.25 => 1)
   end
 
   context 'when the app raises an exception' do
@@ -62,27 +62,34 @@ describe Prometheus::Client::Rack::Collector do
     end
 
     it 'traces exceptions' do
-      labels = { exception: 'NoMethodError' }
-
       expect { get '/broken' }.to raise_error NoMethodError
 
-      expect(registry.get(:http_exceptions_total).get(labels)).to eql(1.0)
+      metric = :http_server_exceptions_total
+      labels = { exception: 'NoMethodError' }
+      expect(registry.get(metric).get(labels)).to eql(1.0)
     end
   end
 
-  context 'setting up with a block' do
+  context 'when using a custom counter label builder' do
     let(:app) do
-      described_class.new(original_app, registry: registry) do |env|
-        { method: env['REQUEST_METHOD'].downcase } # and ignore the path
-      end
+      described_class.new(
+        original_app,
+        registry: registry,
+        counter_label_builder: lambda do |env, code|
+          {
+            code:   code,
+            method: env['REQUEST_METHOD'].downcase,
+          }
+        end,
+      )
     end
 
     it 'allows labels configuration' do
       get '/foo/bar'
 
+      metric = :http_server_requests_total
       labels = { method: 'get', code: '200' }
-
-      expect(registry.get(:http_requests_total).get(labels)).to eql(1.0)
+      expect(registry.get(metric).get(labels)).to eql(1.0)
     end
   end
 end
